Concurrent processing + strictly ordered plugin handlers
Basically, we fetch and process all of the block information
concurrently. Since plugins need a strict ordering guarantee
(events that occur first on chain are seen and processed
before later ones), we track the state of indexing
and concurrently call plugin handlers once we
have the data for a continuous sequence of blocks.